requirements.txt
kivy==2.2.1
Pillow==10.2.0 